Changed references in File and Commented out nonexistent fonts.
Change 1:
The font references should not start with `/fonts/kanit/`, since stylesheet.css is in the same directory as the fonts. This has been rectified.
Change 2:
Most fonts specified in the file actually do not have woff font files corresponding to them. These have been commented out and can be uncommented once the fonts are added.